Move code from protocol.rs to lib.rs
